// SDKs/apideck.com/lead/8.70.2/typescript/src/sdk/models/shared/notfoundresponse.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class NotFoundResponse extends SpeakeasyBase {
  @Metadata({ data: "json, name=detail" })
  detail?: any;

  @Metadata({ data: "json, name=error" })
  error?: string;

  @Metadata({ data: "json, name=message" })
  message?: string;

  @Metadata({ data: "json, name=ref" })
  ref?: string;

  @Metadata({ data: "json, name=status_code" })
  statusCode?: number;

  @Metadata({ data: "json, name=type_name" })
  typeName?: string;
}